requirements.txt
pyserial==3.5
# readline
requests